Plot observed saturation
This IDL script plots the execution time observed while running the TOTAL( ) function within procedure DOUBLESEARCH of procedure GRIDFIT_VEC. The call to TOTAL( ) 
seem to be suitable for multi-threading but is limited to a single CPU. The matrix is passed to TOTAL( ) and evaluated on one CPU. As the matrix grows in size 
the CPU seems to saturate and this is exhibited through a slow down in the execution time. The slow down is illustrated in the plot made via this IDL script.